+++ changelog.md
@@ -1,6 +1,13 @@
 
 ## Changelog for the `groan_rs` library
 
+### Version 0.10.0
+
+#### Changes to `System::traj_iter_map_reduce` (as is tradition)
+- **Breaking change:** The `Data` structure in `System::traj_iter_map_reduce` no longer needs to implement `Add`. Instead, it needs to implement `ParallelTrajData` which requires the user to specify how the data structures should be reduced (merged). `ParallelTrajData` also allows the user to provide a custom `initialize` function which accepts thread ID and is automatically called after spawning a thread. This allows the user to implement thread-specific behavior or to properly sort the final results.
+
+***
+
 ### Version 0.9.0
 
 #### Atom Index